% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epl_00_18.R
\docType{data}
\name{epl_00_18}
\alias{epl_00_18}
\title{Data on English Premier League, 2000-2018:}
\format{A data frame with 6840 rows and 6 variables

Home: The name of the home team

Away: The name of the away team

HG: Full Time Home Team Goals

AG: Full Time Away Team Goals

Round: the number of the round (34 overall)

The Season when the match was played (eg, season 2013/2014 is given as 2013)}
\source{
\href{http://www.worldfootball.net}{WorldFootball.net}
}
\usage{
data(epl_00_18)
}
\description{
Data on English Premier League, 2000-2018:
}
\keyword{Elo}
\keyword{football,}
\keyword{ratings}
