% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoxScore.R
\name{getBoxScore}
\alias{getBoxScore}
\title{NBA regular season box scores}
\usage{
getBoxScore(year)
}
\arguments{
\item{year}{4 digits number of the season year. Fo example, season 2010-2011 will be given as 2010}
}
\value{
dataframe with regular season results
}
\description{
Gets box score for for NBA regular season. Uses teamgamelogs? from stats.nba.com
}
\author{
Vardan Kazaryan
}
