% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPassing.R
\name{getPassing}
\alias{getPassing}
\title{NBA players passing statistics}
\usage{
getPassing(year, playerID, seasonType)
}
\arguments{
\item{year}{The season year, i.e. 2011 for the season 2010-2011}

\item{playerID}{The player ID from NBA database.}

\item{seasonType}{Regular, Playoffs}
}
\value{
dataframe with players passing statistics
}
\description{
Gets players passing statistics for the given season
}

