% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerBio.R
\name{getPlayerBio}
\alias{getPlayerBio}
\title{NBA player's BIO}
\usage{
getPlayerBio(playerID)
}
\arguments{
\item{playerID}{as specified by nba.com}
}
\value{
dataframe with player's bio
}
\description{
Gets player's bio from stats.nba.com
}

