% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPlayerID.R
\name{getPlayerID}
\alias{getPlayerID}
\title{NBA player IDS}
\usage{
getPlayerID(year)
}
\arguments{
\item{year}{The season year, i.e. 2011 for the season 2010-2011}
}
\value{
dataframe with player names, ids, team names and ids
}
\description{
Gets player and team IDs for active roster for the given season
}
