% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTeamRoster.R
\name{getTeamRoster}
\alias{getTeamRoster}
\title{get Team roster}
\usage{
getTeamRoster(Year, TeamID)
}
\arguments{
\item{Year}{The season year, i.e. 2011 for the season 2010-2011}

\item{TeamID}{Team ID}
}
\value{
team roster
}
\description{
Gets team roster by team ID
}
