% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_salary.R
\docType{data}
\name{mlb_salary}
\alias{mlb_salary}
\title{MLB salaries, 1999-2017}
\format{A data frame with 608 rows and 6 variables

Team: team name

Opening.day: Opening day salary

Current: Current Season Salary

Avg: Average salary

Median: Median salary

season: The season}
\source{
\href{https://www.usatoday.com/sports/mlb/salaries/2016/team/all/}{USATODAY}
}
\usage{
data(mlb_salary)
}
\description{
MLB salaries, 1999-2017
}
\keyword{balance}
\keyword{competitive}
\keyword{mlb,}
