% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_st99_16.R
\docType{data}
\name{mlb_st99_16}
\alias{mlb_st99_16}
\title{MLB final standings, 1999-2017, Regular season}
\format{A data frame with 540 rows and 6 variables

Team: team name

Lig: The conference/League - AL: American League, NL: National League

Div: Division: ED-Eastern Division, WD: Western Division, CD: Central Division

W: Number of wins

L: number of lost games

season: The season}
\source{
\href{https://www.baseball-reference.com}{www.baseball-reference.com}
}
\usage{
data(mlb_st99_16)
}
\description{
MLB final standings, 1999-2017, Regular season
}
\keyword{and}
\keyword{balance,}
\keyword{competitive}
\keyword{luck}
\keyword{mlb,}
\keyword{skill}
