% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlb_standings.R
\docType{data}
\name{mlb_standings}
\alias{mlb_standings}
\title{MLB final standings since 1999}
\format{
A data frame with 780 rows and 10 variables

season: The season, 2017 stands for the season 2017-2018

Team: team name

Lig: The conference/League - AL: American League, NL: National League

Div: Division: E-Eastern Division, W: Western Division, C: Central Division

W: Number of wins

L: number of lost games

G: Number of games played

 R: Runs scored

 RA: Runs allowed

 Rank: Rank within the division
}
\source{
\href{https://www.baseball-reference.com}{www.baseball-reference.com}
}
\usage{
data(mlb_standings)
}
\description{
MLB final standings since 1999
}
\keyword{and}
\keyword{balance,}
\keyword{competitive}
\keyword{luck}
\keyword{mlb,}
\keyword{skill}
