% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba12p.R
\docType{data}
\name{nba12p}
\alias{nba12p}
\title{NBA player statistics for 2011-2012 from BasketValue}
\format{A data frame with 479 rows and 39 variables}
\source{
BasketValue
}
\usage{
data(nba12p)
}
\description{
NBA player statistics for 2011-2012 from BasketValue
}
\keyword{Adjusted}
\keyword{NBA,}
\keyword{Plus/Minus}
\keyword{Plus/Minus,}
