% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba2009_2018.R
\docType{data}
\name{nba2009_2018}
\alias{nba2009_2018}
\title{Data on NBA games for regular season: 2009-2018}
\format{A data frame with 12060 rows and 9 variables. Each row is one game

Season_ID: 2009 stands for the season 2009-2010, etc

GAME_DATE

home.TEAM_ABBREVIATION

home.TEAM_NAME

home.PTS: Points scored by home team

away.TEAM_ABBREVIATION

away.TEAM_NAME

away.PTS: Points scored by away team

home.WL: The result of the game -W: home team won, L: home team lost}
\usage{
data(nba2009_2018)
}
\description{
Data on NBA games for regular season: 2009-2018
}
\keyword{nba}
