% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_games.R
\docType{data}
\name{nba_games}
\alias{nba_games}
\title{Data on NBA games for regular season since 2009}
\format{
A data frame with 20349 rows and 9 variables. Each row is one game

Season_ID: 2009 stands for the season 2009-2010, etc

GAME_DATE

home.TEAM_ABBREVIATION

home.TEAM_NAME

home.PTS: Points scored by home team

away.TEAM_ABBREVIATION

away.TEAM_NAME

away.PTS: Points scored by away team

home.WL: The result of the game -W: home team won, L: home team lost
}
\usage{
data(nba_games)
}
\description{
Data on NBA games for regular season since 2009
}
\keyword{nba}
