% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_misc.R
\docType{data}
\name{nba_misc}
\alias{nba_misc}
\title{Miscellaneous data on NBA team statistics, season 2016-2017}
\format{A data frame with 30 rows and 26 variables

For variable definitions, look at the source

variables with suffix _o, are showing offensive statistics

variables with suffix _d, are showing defensive statistics

https://www.basketball-reference.com/leagues/NBA_2017.html#misc_stats}
\usage{
data(nba_misc)
}
\description{
Miscellaneous data on NBA team statistics, season 2016-2017
}
\keyword{nba,}
\keyword{skill_luck}
