% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_players.R
\docType{data}
\name{nba_players}
\alias{nba_players}
\title{NBA Players statistics since 1950}
\format{
A data frame with 24691 rows and 53 variables. Each row is one season for one player.

X: Index

Year: Season

Player: Player's name

Position: Position of the player

Age: Age of the player

Tm: Team of the playe.

G: Games

GS: Games Started

MP: Minutes Played

PER: Player Efficiency Rating

TS: True Shooting perc

3PAr: 3-Point Attempt Rate

FTr: Free Throw Rate

ORBperc: Offensive Rebound Percentage

DRBperc: Defensive Rebound Percentage

TRBperc: Total Rebound Percentage

ASTperc: Assist Percentage

STLperc: Steal Percentage

BLKperc: Block Percentage

TOVperc: Turnover Percentage

USGperc: Usage Percentage

OWS: Offensive Win Shares

DWS: Defensive Win Shares

WS: Win Shares

WS/48: Win Shares Per 48 Minutes

OBPM: Offensive Box Plus/Minus

DBPM: Defensive Box Plus/Minus

BPM: Box Plys/Minus

VORP: Value Over Replacement

FG: Field Goals

FGA: Field Goal Attempts

FGperc: Field Goal Percentage

3P: 3-Point Field Goals

3PA: 3-Point Field Goal Attempts

3Pperc: 3-Point Field Goal Percentage

2P: 2-Point Field Goals

2PA: 2-Point Field Goal Attempts

2Pperc: 2-Point Field Goal Percentage

eFGperc: Effective Field Goal Percentage

FT: Free Throws

FTA: Free Throw Attempts

FTperc: Free Throw Percentage

ORB: Offensive Rebounds

DRB: Defensive Rebounds

TRB: Total Rebounds

AST: Assists

STEALS: Steals

BLK: Blocks

TOV: Turnovers

PF: Personal Fouls

PTS: Points
}
\usage{
data(nba_players)
}
\description{
NBA Players statistics since 1950
}
\keyword{nba_players}
