% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl.R
\docType{data}
\name{nfl}
\alias{nfl}
\title{Data on NFL matches/regular season}
\format{
A data frame with 6712 rows and 13 variables

schedule_season: from 2000-2022

schedule_week: The number of the week when the game was played

schedule_date: The actual date of the game

team_away: Away team

score_away: The score of the away team

team_home: Home team

score_home: The score of home team

stadium: The name of the stadium

stadium_type: indoor, outdoor, retractable

stadium_location: The location of the stadium: City and state abbreviation

result: The result of the game: H=Home Win, T=Tie, A=Away Win

spread: points spread, the difference between away team and home team scores

home_odds_open, home_odds_close: Opening and closing decimal odds for the home team

away_odds_open, away_odds_open: Opening and closing decimal odds for the away team (The data is missing for some seasons)

Source for odds: http://www.aussportsbetting.com/data/historical-nfl-results-and-odds-data/
}
\usage{
data(nfl)
}
\description{
Data on NFL matches/regular season
}
\keyword{Elo}
\keyword{betting}
\keyword{nfl,}
\keyword{rating,}
\keyword{strategy}
