% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_st.R
\docType{data}
\name{nfl_st}
\alias{nfl_st}
\title{NFL final standings, 1998-2018, Regular season}
\format{A data frame with 667 rows and 16 variables

Tm: team name

W: Number of wins

L: number of lost games

T: Number of Tied games

Wpct: Winning percentage

PF: Points Scored by team

PA: Points Scored by opposition


MoV: Margin of Victory (Points Scored - Points Allowed)/ Games Played

SoS: Strength of Schedule Average quality of opponent as
measured by SRS (Simple Rating System)

SRS:  (Simple Rating System) SRS = MoV + SoS = OSRS + DSRS

OSRS: Offensive SRS

DSRS: Defensive SRS

Season: Season

Conf: Conference

Div: Division}
\source{
\href{https://www.pro-football-reference.com}{Pro football reference}
}
\usage{
data(nfl_st)
}
\description{
NFL final standings, 1998-2018, Regular season
}
\keyword{and}
\keyword{balance,}
\keyword{competitive}
\keyword{luck}
\keyword{nfl,}
\keyword{skill}
