% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl_st99_16.R
\docType{data}
\name{nfl_st99_16}
\alias{nfl_st99_16}
\title{NFL final standings, 1999-2016, Regular season}
\format{A data frame with 573 rows and 7 variables

Team: team name

Conf: The conference

Div: Division

W: Number of wins

L: number of lost games

T: Number of Tied games

season: The season}
\source{
\href{http://www.nfl.com/standings}{nfl.com}
}
\usage{
data(nfl_st99_16)
}
\description{
NFL final standings, 1999-2016, Regular season
}
\keyword{and}
\keyword{balance,}
\keyword{competitive}
\keyword{luck}
\keyword{nfl,}
\keyword{skill}

