% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriea_st.R
\docType{data}
\name{seriea_st}
\alias{seriea_st}
\title{Data final standings for Serie A - 1996-2017}
\format{A data frame with 422 rows and 10 variables

Rank- Team's rank at the end of the season

Team - Team's name

M - Matches played

W - Number of wins

D - number of draws

L- number of loses

Goals: Goals for:Goals against

Diff: Goal difference Goals for minus Goals against

Pt: Points earned at the end of the season

Season: The end year of the season (season 2011/2012 is recorded as 2012)}
\source{
\href{http://www.worldfootball.net}{World Football data}
}
\usage{
data(seriea_st)
}
\description{
Data final standings for Serie A - 1996-2017
}
\keyword{A}
\keyword{Serie}
\keyword{football,}
