% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfers.R
\docType{data}
\name{transfers}
\alias{transfers}
\title{Data on football transfers. Each row is one transfer}
\format{
A data frame with 6248 rows and 10 variables

Note: The dataset might be incomplete with some minor errors.
Any help in updating or cleaning the dataset will be appreciated.

PLAYER: The name of the player

WINDOW: The transfer windows: "Pre-Season", "Mid-Season", NA

POSITION: Position of the player "Midfielder", "Attacker", "Defender", "Goalkeeper"

COUNTRY: Players home country

FROM: The team from which the player was transfered

TO: The team to which the player was transfered

DESCRIPTION: The description of the transfer deal - Sold, Loan, etc

PRICE: The price of the transfer

LEAGUE: The league where the player moved

SEASON: The transfer season
}
\usage{
data(transfers)
}
\description{
Data on football transfers. Each row is one transfer
}
\keyword{football}
