% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_14_final.R
\docType{data}
\name{wc_14_final}
\alias{wc_14_final}
\title{Passing data from FIFA WC 2014 Final}
\format{A list with 4 dataframes inside

argentina_team, germany_team: players aggregated passing statistics

The first 11 rows are the starting line-up of the teams

argentina_passing, germany_passing: Weighted sociomatrix of the passes made.
The row/columns names are the shirt numbers of the players.}
\source{
\href{http://www.fifa.com/worldcup/matches/round=255959/match=300186501/index.html}{fifa.com}
}
\usage{
data(wc_14_final)
}
\description{
Passing data from FIFA WC 2014 Final
}
\keyword{analysis}
\keyword{network}
\keyword{soccer,}

