% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_table.R
\name{final_table}
\alias{final_table}
\title{Final League table}
\usage{
final_table(data, country, season)
}
\arguments{
\item{data}{dataframe with the same structure as f_data_sm}

\item{country}{The country for which the table is created}

\item{season}{the season for which the table is returned, season 2017/2018 will be given as 2018}
}
\value{
dataframe with the final league standing
}
\description{
Creates final league table
}
