% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_east.R
\docType{data}
\name{nba_east}
\alias{nba_east}
\title{Data on NBA final standings, Eastern Conference, 1991-2017}
\format{A data frame with 387 rows and 7 variables

Rank: The final standing

Team: Team name

W: Number of Wins

L: Number of loses

Pct: Winning percentage

GB:  Games Behind

Season: The end year of the season (season 2011/2012 is recorded as 2012)}
\source{
\href{http://www.landofbasketball.com}{Land of Basketball}
}
\usage{
data(nba_east)
}
\description{
Data on NBA final standings, Eastern Conference, 1991-2017
}
\keyword{NBA}

