% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nba_west.R
\docType{data}
\name{nba_west}
\alias{nba_west}
\title{Data on NBA final standings, Western Conference, 1991-2018}
\format{A data frame with 402 rows and 9 variables

Rank: The final standing

Team: Team name

W: Number of Wins

L: Number of loses

Pct: Winning percentage

GB:  Games Behind

PS: Points Scored

PA: Points Allowed

Season: The end year of the season (season 2011/2012 is recorded as 2012)}
\source{
\href{http://www.landofbasketball.com}{Land of Basketball}
}
\usage{
data(nba_west)
}
\description{
Data on NBA final standings, Western Conference, 1991-2018
}
\keyword{NBA}
